fix: align experience status sync with client

+++ src/Godswar.Server/Packets/ClientStatusEffect.cs
@@ -2,6 +2,6 @@ namespace Godswar.Server.Packets;
 
 /// <summary>
 /// A status entry rendered by the original client status bar.
-/// Remaining time is encoded as the original protocol's unsigned 16-bit seconds.
+/// Remaining time is encoded as this client revision's unsigned 32-bit seconds.
 /// </summary>
-internal readonly record struct ClientStatusEffect(uint StatusId, ushort RemainingSeconds);
+internal readonly record struct ClientStatusEffect(uint StatusId, uint RemainingSeconds);
